Refactoring and Optimization
